chore: remove extra contact-page js file
